back/pyproject.toml
[project]
name = "back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "google-genai>=1.15.0",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "uvicorn[standard]>=0.34.2",
]
